+++ .eslintrc.js
@@ -4,7 +4,7 @@ module.exports = {
     node: true,
     "vue/setup-compiler-macros": true,
   },
-  ignorePatterns: ["dist/**/*"],
+  ignorePatterns: ["dist/**/*", "trees/**/*", "types/**/*"],
   extends: [
     "eslint:recommended",
     "plugin:vue/vue3-essential",
