+++ dev/tsconfig.json
@@ -1,3 +1,9 @@
 {
   "extends": "../tsconfig",
+  "include": [
+    "./**/*.ts",
+    "./**/*.d.ts",
+    "./**/*.tsx",
+    "./**/*.vue"
+  ]
 }
