+++ src/composables/date.ts
@@ -1,0 +1,4 @@
+export interface DateRange {
+  minDate: number
+  maxDate: number
+}
