+++ src/composables/overlay.ts
@@ -1,0 +1,4 @@
+export interface Flash {
+  type?: string
+  message: string
+}
