+++ src/lib-components/layout/SidebarLayout.vue
@@ -13,7 +13,7 @@ import {
 } from "@headlessui/vue"
 import { MenuAlt2Icon, XIcon } from "@heroicons/vue/outline"
 import { CogIcon } from "@heroicons/vue/solid"
-import * as NavTypes from "../../types/nav"
+import * as NavTypes from "@/composables/nav"
 import { ref } from "vue"
 
 const props = withDefaults(
