+++ src/lib-components/lists/StaticTable.vue
@@ -1,6 +1,6 @@
 <script setup lang="ts">
 import { ComponentPublicInstance, getCurrentInstance } from "vue"
-import * as TableTypes from "../../types/table"
+import * as TableTypes from "@/composables/table"
 
 defineProps<{
   tableData: TableTypes.Static
