+++ src/tsconfig.json
@@ -2,7 +2,6 @@
   "extends": "../tsconfig",
   "compilerOptions": {
     "declaration": true,
-    "emitDeclarationOnly": true,
     "outDir": "../types"
   },
   "include": [
