--- src/types/components.d.ts
@@ -1,8 +1,0 @@
-import { TreesComponents } from "@/lib-components"
-
-export default TreesComponents
-
-declare module "@vue/runtime-core" {
-  // eslint-disable-next-line @typescript-eslint/no-empty-interface
-  interface GlobalComponents extends TreesComponents {}
-}
