--- src/types/global.d.ts
@@ -1,10 +1,0 @@
-import { Emitter } from "mitt"
-
-export {}
-
-declare global {
-  interface Window {
-    Flashes: Array<{ type?: string; message: string }>
-    VueBus: Emitter
-  }
-}
