--- src/types/index.d.ts
@@ -1,3 +1,0 @@
-export * as NavTypes from "./nav"
-export * as TableTypes from "./table"
-export * as UserTypes from "./users"
